+++ FibonnacciSequence/src/timing/FibonacciTimingDecorator.java
@@ -13,9 +13,9 @@ public class FibonacciTimingDecorator implements FibonacciInterface {
     }
 
     @Override
-    public int fibonacci(int n) {
+    public long fibonacci(int n) {
         long startTime = System.nanoTime();
-        int result = fibonacciMethod.fibonacci(n);
+        long result = fibonacciMethod.fibonacci(n);
         long endTime = System.nanoTime();
         long executionTime = endTime - startTime;
